Zadanie 3.6: testy jednostkowe dla kalkulatora

+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/Calculator.java
@@ -1,14 +1,14 @@
 package com.kodilla.basic_assertion;
 
 public class Calculator {
-    public int sum (int a, int b) {
-        return a + b;
+    public double sum (double a, double b) {
+        return (a + b);
 }
-    public int sub (int a, int b) {
+    public double sub (double a, double b) {
         return a - b;
 
     }
-    public int squ (int a) {
+    public double squ(double a) {
         return a * a;
     }
 }
